+++ frontend/src/plugins/demo/guide.ts
