+++ frontend/src/plugins/demo/hint.ts
@@ -1,0 +1,2 @@
+// Hint is a special guide that has no Next button and is always shown.
+export default null;
