+++ frontend/src/plugins/demo/types.ts
