+++ frontend/static/demo/schema-migration/recorder.json
